update live training workflow
